+++ utils/types/jec-commons-node.d.ts
@@ -7,6 +7,10 @@
 
 declare module "jec-commons-node" {
 
+/// <reference types="node" />
+import { FileProperties, FileStats, PathStats } from "jec-commons";
+import * as fs from "fs";
+
 /*PLACEHOLDER*/
 
 }