+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# JEC Node.js Commons Project: Update Release Notes
+
+<a name="jec-commons-node-1.0.1"></a>
+## **1.0.1** (2017-06-05)
+
+### Bug Fixes
+
+### Features
+
+- **FileStatsProxy:** an utility class for the proxifying the `fs.Stats` API
+
+- Dependencies upgrade
+
 <a name="jec-commons-node-1.0.0"></a>
 ## **1.0.0** (2017-06-03)